*** THIS BREAKS THE CRON JOBS *** we need to call the shell scripts now
Updated CakePHP, fixed gitignore for images;

+++ app/Model/ReviewFrame.php
@@ -72,9 +72,9 @@ class ReviewFrame extends AppModel
 		$dataT2 = $this->query("SELECT SUM(liking_qty) as qty FROM (SELECT count(*) AS liking_qty, AVG(groove) AS avg_groove FROM review_frames AS ReviewFrames WHERE $condition group by position HAVING avg_groove > .75) as t2");
 		$dataT3 = $this->query("SELECT SUM(disliking_qty) as qty FROM (SELECT count(*) AS disliking_qty, AVG(groove) AS avg_groove FROM review_frames AS ReviewFrames WHERE $condition group by position HAVING avg_groove < .25 && avg_groove > 0) as t3");
 
-        $liking     = Hash::get($dataT2, "0.0.qty");
-        $disliking  = Hash::get($dataT3, "0.0.qty");
-        $total      = Hash::get($dataT1, "0.0.qty");
+        $liking     = (int)Hash::get($dataT2, "0.0.qty");
+        $disliking  = (int)Hash::get($dataT3, "0.0.qty");
+        $total      = (int)Hash::get($dataT1, "0.0.qty");
         $neutral    = $total - $disliking - $liking;
 
         // this prevents divisions by 0
